Update what-is.md
On line #5 "was built" should be changed to "is built". Present tense is usually a little better. For example, on the S3 page -https://docs.aws.amazon.com/AmazonS3/latest/userguide/Welcome.html  The phrase "Amazon S3 is intentionally built" is used. 

Thank you for taking the time to review my request. 

-Brian